Refactor to delete methods made obsolete by SchemaAnalyzer
